Added Win32GameControl.
